Replace TempDir with temp option on Dir creation

+++ src/Dir.ts
@@ -3,20 +3,28 @@ import * as path from 'node:path';
 import sanitizeFilename from 'sanitize-filename';
 import { File } from './File.ts';
 
+export type DirOptions = {
+  temp?: boolean;
+};
+
 /**
  * Reference to a specific directory with methods to create and list files.
- * Default path: '.'
- * > Created on file system the first time .path is read or any methods are used
+ * @param inputPath
+ * The path of the directory, created on file system the first time `.path` is read or any methods are used
+ * @param options
+ * include `{ temp: true }` to enable the `.clear()` method
  */
 export class Dir {
   #inputPath;
   #resolved?: string;
+  isTemp;
 
   /**
    * @param path can be relative to workspace or absolute
    */
-  constructor(inputPath = '.') {
+  constructor(inputPath: string, options: DirOptions = {}) {
     this.#inputPath = inputPath;
+    this.isTemp = Boolean(options.temp);
   }
 
   /**
@@ -32,23 +40,26 @@ export class Dir {
 
   /**
    * Create a new Dir inside the current Dir
-   * @param subPath joined with parent Dir's path to make new Dir
+   * @param subPath
+   * joined with parent Dir's path to make new Dir
+   * @param options
+   * include `{ temp: true }` to enable the `.clear()` method
    * @example
    * const folder = new Dir('example');
    * // folder.path = '/path/to/cwd/example'
    * const child = folder.dir('path/to/dir');
    * // child.path = '/path/to/cwd/example/path/to/dir'
    */
-  dir(subPath: string) {
-    return new Dir(path.join(this.path, subPath));
+  dir(subPath: string, options: DirOptions = {}) {
+    return new Dir(path.join(this.path, subPath), options);
   }
 
   /**
-   * Creates a new TempDir inside current Dir
+   * Creates a new temp directory inside current Dir
    * @param subPath joined with parent Dir's path to make new TempDir
    */
   tempDir(subPath: string) {
-    return new TempDir(path.join(this.path, subPath));
+    return this.dir(subPath, { temp: true });
   }
 
   sanitize(filename: string) {
@@ -74,21 +85,9 @@ export class Dir {
   get files() {
     return fs.readdirSync(this.path).map(filename => this.file(filename));
   }
-}
 
-/**
- * Extends Dir class with method to `clear()` contents.
- * Default path: `./.temp`
- */
-export class TempDir extends Dir {
-  constructor(inputPath = `./.temp`) {
-    super(inputPath);
-  }
-
-  /**
-   * > ⚠️ Warning! This deletes the directory!
-   */
   clear() {
+    if (!this.isTemp) throw new Error('Dir is not temporary');
     fs.rmSync(this.path, { recursive: true, force: true });
     fs.mkdirSync(this.path, { recursive: true });
   }
